chore: move to ee dir
