+++ client/src/components/ee/LICENSE
@@ -1,0 +1,9 @@
+# Enterprise Edition License
+
+Version 0.0
+
+## Terms and Conditions
+
+Commercial license coming soon. Details to follow when available.
+
+Until that time, no third party has any right to use any contents or code of any `/ee` directory.
